Diagonals working, lurch cause pinpointed
- The diagonals needed to have each key be a modifier of the
other. e.g. if "w" is pressed and then "a" is pressed, "w" needs
to be the modifier or "a", but if "a" is pressed first, it needs
to be made the modifier of "w", otherwise, in the second case,
"w" would be taken as the last input and the object would only
move up, not sideways.
- The lurch happens because force is added on the key event. When
a key is held down, there's a little pause before it starts
repeating, so the object gets pushed a bit, pauses, then gets
pushed consistently.